moviebot/requirements.txt
# ====================================
# 🎬 MovieBot - Core Dependencies
# ====================================

# === 🤖 Bot Frameworks ===
discord.py==2.3.2                    # Discord API wrapper
python-telegram-bot==20.6            # Telegram bot framework

# === 🎞️ Media Processing ===
ffmpeg-python==0.2.0                 # Python bindings for FFmpeg
yt-dlp==2024.04.09                   # YouTube video downloader (ffmpeg required)

# === ⚙️ System / File Handling ===
aiofiles==23.2.1                     # Async file I/O
psutil==5.9.8                        # Process and system monitoring
filelock==3.13.4                     # File-based locking for cross-process safety

# === 🧪 Config & Environment ===
python-dotenv==1.0.1                 # Load env vars from `.env`

# === 🧾 Logging ===
loguru==0.7.2                        # Elegant structured logger

# === 🌐 Async + Web Client (optional but recommended) ===
aiohttp==3.9.5                       # Async HTTP client (for future APIs or integrations)
asyncio==3.4.3                       # For legacy use; skip if using Python 3.7+

# === ⚡ Optional: Web Panel / API Support ===
fastapi==0.111.0                     # Optional web control/API panel
uvicorn==0.29.0                      # ASGI server to run FastAPI
pydantic==2.7.3                      # Schema validation for FastAPI

# === 🐧 Linux Display / GTK (optional) ===
PyGObject==3.46.0                    # For GTK/X11 integration, if needed

# === 🧪 Dev Tools (optional) ===
pytest==8.2.1                        # Unit testing
black==24.4.2                        # Code formatting

# ====================================
# 🔐 NOTES:
# - Don't upgrade versions blindly. Breakage may occur in ffmpeg, ptb, or yt-dlp.
# - Consider using a `venv` or Docker to isolate versions.
# - Use: `pip install -r requirements.txt`
# ====================================
